backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
anyio==4.11.0
    # via httpx
asgiref==3.10.0
    # via
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.10.2
    # via daphne
automat==25.4.16
    # via twisted
billiard==4.2.2
    # via celery
boto3==1.40.69
    # via -r requirements.in
botocore==1.40.69
    # via
    #   boto3
    #   s3transfer
celery==5.5.3
    # via -r requirements.in
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   sendgrid
    #   service-identity
daphne==4.2.1
    # via -r requirements.in
dj-database-url==3.0.1
    # via -r requirements.in
django==5.2.8
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via twisted
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
markupsafe==3.0.3
    # via werkzeug
packaging==25.0
    # via
    #   gunicorn
    #   kombu
pillow==12.0.0
    # via qrcode
prompt-toolkit==3.0.52
    # via click-repl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.3.0
    # via twisted
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
python-dotenv==1.2.1
    # via -r requirements.in
python-http-client==3.3.7
    # via sendgrid
qrcode[pil]==8.2
    # via -r requirements.in
redis==7.0.1
    # via -r requirements.in
s3transfer==0.14.0
    # via boto3
sendgrid==6.12.5
    # via -r requirements.in
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via django
twisted[tls]==25.5.0
    # via daphne
txaio==25.9.2
    # via autobahn
typing-extensions==4.15.0
    # via twisted
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
werkzeug==3.1.3
    # via sendgrid
zope-interface==8.0.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
